requirements.txt
Flask ; ==0.12.2
numpy ; ==1.14.2
pypng ; ==0.0.18
